requirements.txt
alembic==1.11.3
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==40.0.2
distlib==0.3.6
dnspython==2.4.2
email-validator==2.0.0.post2
filelock==3.12.2
Flask==2.2.5
Flask-Cors==4.0.0
Flask-Login==0.6.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
future==0.18.3
gevent==23.7.0
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==7.0.1
is-disposable-email==1.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
packaging==23.1
platformdirs==3.9.1
psycopg2-binary==2.9.9
pycparser==2.21
PySocks==1.7.1
python-dotenv==0.21.1
python-dotenv-vault==0.6.0
requests==2.31.0
requests-file==1.5.1
six==1.16.0
SQLAlchemy==2.0.19
tldextract==3.4.4
typing_extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
WTForms==3.0.1
wtforms-validators==1.0.0
zipp==3.17.0
zope.event==5.0
zope.interface==6.0
